requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.3
boto3==1.35.0
botocore==1.35.99
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.2
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.2
cssselect2==0.8.0
daphne==4.2.1
decorator==5.2.1
Django==5.2.6
django-celery-results==2.5.1
django-cors-headers==4.9.0
django-ninja==1.4.3
django-storages==1.14.4
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebase-admin==6.5.0
fonttools==4.60.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
httplib2==0.31.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
jmespath==1.0.1
kombu==5.6.1
moviepy==2.2.1
msgpack==1.1.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
reportlab==4.4.5
requests==2.32.5
rsa==4.9.1
s3transfer==0.10.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zope.interface==8.0.1
zopfli==0.4.0
